requirements.txt
# Core dependencies
numpy>=1.24.0
mss>=9.0.0
pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
treys>=0.1.8

# Tkinter is built-in with Python on Windows

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
